dom/bindings/test/test_exceptions_from_jsimplemented.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=923010
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 923010</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">
  /** Test for Bug 923010 **/
  try {
    var conn = new RTCPeerConnection();

    var candidate = new RTCIceCandidate({candidate: "x" });
    conn.addIceCandidate(candidate)
    .then(function() {
      ok(false, "addIceCandidate succeeded when it should have failed");
    }, function(reason) {
      is(reason.lineNumber, 17, "Rejection should have been on line 17");
      is(reason.message,
         "Invalid candidate (both sdpMid and sdpMLineIndex are null).",
         "Should have the rejection we expect");
    })
    .catch(function(reason) {
      ok(false, "unexpected error: " + reason);
    });
  } catch (e) {
    // b2g has no WebRTC, apparently
    todo(false, "No WebRTC on b2g yet");
  }

  conn.close();
  try {
      conn.setIdentityProvider("example.com", { protocol: "foo" });
      ok(false, "That call to setIdentityProvider should have thrown");
    } catch (e) {
      is(e.lineNumber, 36, "Exception should have been on line 36");
      is(e.message,
         "Peer connection is closed",
         "Should have the exception we expect");
    }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=923010">Mozilla Bug 923010</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
